refactor: slick의 중복 렌더링을 방지하기 위한 스켈레톤 처리 (#88)
